[UPD] README.rst, manifest and files refactoring
