+++ project_phase/__manifest__.py
@@ -3,11 +3,12 @@
 
 {
     "name": "Project Phase",
-    "version": "11.0.1.1.0",
+    "version": "11.0.1.0.0",
     "category": "Project",
     "license": "AGPL-3",
-    "author": "AvanzOSC",
-    "website": "http://www.avanzosc.es",
+    "author": "AvanzOSC,"
+              "Odoo Community Association (OCA)",
+    "website": "https://github.com/OCA/project",
     "contributors": [
         "Oihane Crucelaegui <oihanecrucelaegi@avanzosc.es>",
         "Ana Juaristi <anajuaristi@avanzosc.es>",
