+++ project_phase/models/__init__.py
@@ -1,4 +1,5 @@
 # Copyright 2018 Oihane Crucelaegui - AvanzOSC
 # License AGPL-3 - See http://www.gnu.org/licenses/agpl-3.0.html
 
+from . import project_phase
 from . import project_project
