+++ project_phase/readme/CONTRIBUTORS.rst
@@ -1,0 +1,2 @@
+* Oihane Crucelaegui <oihanecrucelaegi@avanzosc.es>
+* Ana Juaristi <anajuaristi@avanzosc.es>
