+++ project_phase/readme/DESCRIPTION.rst
@@ -1,0 +1,1 @@
+With this module you can define different project phases or stages.
